ClauseWise2/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
# PDF & DOCX support
PyPDF2>=3.0.0
python-docx>=1.0.0
# HTTP & testing
requests==2.31.0
pytest==7.4.3
httpx==0.25.2
# Google GenAI SDK (replaces google-cloud-aiplatform)
google-genai>=0.3.0
# Optional: Google Cloud Storage (if you’re uploading files there)
google-cloud-storage==2.10.0
